TP#18772: Changed int.MinValue to long.MinValue

+++ Code/ShipWorks/Stores/Platforms/GenericModule/GenericOrderIdentifier.cs
@@ -43,7 +43,7 @@ namespace ShipWorks.Stores.Platforms.GenericModule
         }
 
         public GenericOrderIdentifier(string orderNumbercomplete)
-            : base(int.MinValue)
+            : base(long.MinValue)
         {
             this.orderNumberComplete = orderNumbercomplete;
         }
